Docker.API.Client

Documentation

dockerAPIVersion :: String

Docker Engine API version. This value will prefix all docker api url paths.

newtype OS

Alias for the system type returned by System.Info.os

Constructors

OS String

defaultDockerUnixSocket :: FilePath

Default docker socket path on unix systems

newDefaultDockerManager :: OS -> IO Manager

Creates a new HTTP connection manager for the default docker daemon address on your system.

newUnixDomainSocketManager :: FilePath -> IO Manager

Creates a new http connection manager from a file path to a unix socket

data DockerClientError

This type describes common errors the docker client might encounter

Constructors

ContainerCreationFailedError String
The request to create a new container failed

UnrecognizedJSONResponseError String
The docker engine API responded with something we didn't expect

GetContainerArchiveError String
The docker engine API responded with an error status code when getting a container archive

NonZeroExitCode String
The container exited with a nonzero exit code

GetContainerLogsError String
The docker engine API responded with an error when we attempted to get a container's logs

ImagePullError String
The docker engine API responded with an error when we attempted to pull an image

type ClientErrorMonad a = ExceptT DockerClientError IO a

Wrapper for composing operations which return an Either DockerClientError a

data ContainerSpec

Describes a docker container to be created with runContainer. Use defaultContainerSpec to get a default value.

Constructors

ContainerSpec

Fields

image :: Text

The image name with an optional tag or digest field (e.g. "python:3.6")

cmd :: [Text]

The container's command. If empty will default to the image default.

user :: Text

Optional user ID to use in the container. If empty will default to the image default.

workingDir :: Text

Optional working directory in the container. If empty will default to the image default.

envVars :: [Text]

Optional list of environment variables of format "FOO=BAR".

hostVolumes :: [Text]

Optional list of host volumes to mount to the container as bind mounts. Must follow the format specied here: https://docs.docker.com/storage/bind-mounts/

defaultContainerSpec

Arguments

:: Text
The container's Docker Image. May optionally include a tag or digest field (e.g. "python:3.6").

-> ContainerSpec

Constructs a simple default ContainerSpec for a docker image which uses the image's default values for all other aguments.

runContainer

Arguments

:: Manager
The connection manager for the docker daemon. You can newDefaultDockerManager to get a default connection manager based on your operating system.

-> ContainerSpec

-> ClientErrorMonad ContainerId

Analagous to the `docker run` command. Runs a container using the input HTTP connection manager and waits until it exits. For more complex use cases you can also use the individual actions that comprise this function. Note that this currently always tries to pull the container's image.

saveContainerArchive

Arguments

:: Manager
The connection manager for the docker daemon

-> UserID
The user id to use for output files and directories

-> GroupID
The group id to use for output files and directories

-> FilePath
The path in the container to copy

-> FilePath
The output path at which to write outputs

-> ContainerId

-> ClientErrorMonad ()

Analagous to the `docker cp` command. Recursively copies contents at the specified path in the container to the provided output path on the host, setting file permissions to the specified user and group id. Note that the container must have been started for this to succeed (i.e. it must have a running or finished state). This method uses conduit to optimize memory usage.

removeContainer

Arguments

:: Manager
The connection manager for the docker daemon

-> Bool
Enable force removal?

-> Bool
Also remove container's volumes?

-> ContainerId

-> ClientErrorMonad ContainerId

Remove a container, equivalent to the `docker container rm` command

saveContainerLogs

Arguments

:: Manager
The connection manager for the docker daemon

-> ContainerLogType
Which logs to fetch from the container

-> FilePath
Output file at which to write logs

-> ContainerId

-> ClientErrorMonad ()

Streams the logs from a docker container into the specified output file path. Logs can include stdout, stderr, or both. Note that if you include both streams, the sorting of the timestamps in the output file may not be perfectly sorted since the stream returned by the docker api is only sorted within each stream type (i.e. stdout and stderr are sorted separately).

data ContainerLogType

Constructors

Stdout

StdErr

Both

pullImage

Arguments

:: Manager
The connection manager for the docker daemon

-> Text
The image of interest. May include an optional tag or digest field.

-> ClientErrorMonad ()

Pulls an image from a remote registry (similar to a `docker pull` command). This currently only supports public registries (e.g. DockerHub).