Docker.API.Client.Internal.Connection

Description

HTTP Connection managers and utilities for connecting to the Docker Engine API

Documentation

defaultDockerUnixSocket :: FilePath

Default docker socket path on unix systems

newDefaultDockerManager :: OS -> IO Manager

Creates a new HTTP connection manager for the default docker daemon address on your system.

newUnixDomainSocketManager :: FilePath -> IO Manager

Creates a new http connection manager from a file path to a unix socket