Docker.API.Client.Internal.Schemas

Description

Aeson schemas for communicating with the Docker Engine API

Documentation

dockerParseOptions :: Int -> Options

Constructs default Aeson options for serializing and deserializing objects in this module. This method drops the first nPrefix characters from record accessors in your Haskell record before serializing to JSON and uses the shortened field names to match record accessors when deserializing.

type ContainerId = String

A Docker container's unique ID (in longer format)

data CreateContainer

Constructors

CreateContainer

Fields

createContainerUser :: Maybe Text

createContainerEnv :: Maybe [Text]

createContainerCmd :: Maybe [Text]

createContainerImage :: Text

createContainerWorkingDir :: Maybe Text

createContainerHostConfig :: Maybe HostConfig

data HostConfig

Constructors

HostConfig

Fields

hostConfigBinds :: Maybe [Text]

data CreateContainerResponse

Constructors

CreateContainerResponse

Fields

createContainerResponseId :: ContainerId

createContainerResponseWarnings :: Maybe [Text]

data WaitContainerResponse

Constructors

WaitContainerResponse

Fields

waitContainerResponseStatusCode :: Int

waitContainerResponseError :: Maybe WaitContainerError

data WaitContainerError

Constructors

WaitContainerError

Fields

waitContainerErrorMessage :: Text