Docker.API.Client.Internal.Types

Documentation

newtype OS

Alias for the system type returned by System.Info.os

Constructors

OS String

data DockerClientError

This type describes common errors the docker client might encounter

Constructors

ContainerCreationFailedError String
The request to create a new container failed

UnrecognizedJSONResponseError String
The docker engine API responded with something we didn't expect

GetContainerArchiveError String
The docker engine API responded with an error status code when getting a container archive

NonZeroExitCode String
The container exited with a nonzero exit code

GetContainerLogsError String
The docker engine API responded with an error when we attempted to get a container's logs

ImagePullError String
The docker engine API responded with an error when we attempted to pull an image

type ClientErrorMonad a = ExceptT DockerClientError IO a

Wrapper for composing operations which return an Either DockerClientError a

data ContainerSpec

Describes a docker container to be created with runContainer. Use defaultContainerSpec to get a default value.

Constructors

ContainerSpec

Fields

image :: Text

The image name with an optional tag or digest field (e.g. "python:3.6")

cmd :: [Text]

The container's command. If empty will default to the image default.

user :: Text

Optional user ID to use in the container. If empty will default to the image default.

workingDir :: Text

Optional working directory in the container. If empty will default to the image default.

envVars :: [Text]

Optional list of environment variables of format "FOO=BAR".

hostVolumes :: [Text]

Optional list of host volumes to mount to the container as bind mounts. Must follow the format specied here: https://docs.docker.com/storage/bind-mounts/

defaultContainerSpec

Arguments

:: Text
The container's Docker Image. May optionally include a tag or digest field (e.g. "python:3.6"). If neither a tag or digest is specified the image will default to the latest tag.

-> ContainerSpec

Constructs a simple default ContainerSpec for a docker image which uses the image's default values for all other aguments.

data ContainerLogType

Constructors

Stdout

StdErr

Both

data DockerStreamType

Possible types of streams returned by the Docker Engine API attach and logs endpoints. Used only internally for parsing the stream metadata returned by docker.

Constructors

DockerStreamStdIn

DockerStreamStdOut

DockerStreamStdErr